@@ -1,7 +1,6 @@
 ﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
-
 using System.Net;
 using System.Text.Json;
 using Microsoft.Azure.Functions.Worker;
@@ -11,7 +10,6 @@ namespace FunctionApp
 {
     public static class Function1
     {
-        //<docsnippet_multiple_outputs>
         [Function("Function1")]
         public static MyOutputType Run(
             [HttpTrigger(AuthorizationLevel.Anonymous, "get", "post", Route = null)] HttpRequestData req,
@@ -22,23 +20,22 @@ namespace FunctionApp
             var response = req.CreateResponse(HttpStatusCode.OK);
 
             response.Headers.Add("Date", "Mon, 18 Jul 2016 16:06:00 GMT");
-            response.Headers.Add("Content", "Content - Type: text / html; charset = utf - 8");
+            response.Headers.Add("Content-Type", "text/html; charset=utf-8");
             response.WriteString("Book Sent to Queue!");
 
             return new MyOutputType()
             {
                 Book = bookVal,
-                HttpReponse = response
+                HttpResponse = response
             };
-        }
-        //</docsnippet_multiple_outputs>
+        }        
 
         public class MyOutputType
         {
             [QueueOutput("functionstesting2", Connection = "AzureWebJobsStorage")]
             public Book Book { get; set; }
 
-            public HttpResponseData HttpReponse { get; set; }
+            public HttpResponseData HttpResponse { get; set; }
         }
 
         public class Book
