@@ -45,8 +45,7 @@ namespace FunctionApp
             var response = httpRequest.CreateResponse(HttpStatusCode.OK);
            
             response.Headers.Add("Date", "Mon, 18 Jul 2016 16:06:00 GMT");
-            response.Headers.Add("Content", "Content - Type: text / html; charset = utf - 8");
-
+            response.Headers.Add("Content-Type", "text/html; charset=utf-8");
             response.WriteString("Welcome to .NET 5!!");
 
             return response;
