@@ -1,7 +1,9 @@
 ﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
+using System;
 using System.Linq;
+using System.Threading.Tasks;
 using Microsoft.Azure.Functions.Worker;
 using Microsoft.Azure.Functions.Worker.Middleware;
 using Microsoft.Azure.Functions.Worker.OutputBindings;
@@ -73,15 +75,13 @@ namespace Microsoft.Extensions.Hosting
                 return builder;
             }
 
-            builder.Services.AddSingleton<OutputBindingsMiddleware>();
+            builder.Services.AddSingleton<OutputBindingsMiddleware>(p => throw new InvalidOperationException($"Type {nameof(OutputBindingsMiddleware)} cannot be resolved using the container."));
 
             builder.Use(next =>
             {
                 return context =>
                 {
-                    var middleware = context.InstanceServices.GetRequiredService<OutputBindingsMiddleware>();
-
-                    return middleware.Invoke(context, next);
+                    return OutputBindingsMiddleware.Invoke(context, next);
                 };
             });
 
@@ -89,7 +89,7 @@ namespace Microsoft.Extensions.Hosting
         }
 
         /// <summary>
-        /// Configures the <see cref="IFunctionsWorkerApplicationBuilder"/> to use provided middleware type.
+        /// Configures the <see cref="IFunctionsWorkerApplicationBuilder"/> to use the provided middleware type.
         /// </summary>
         /// <param name="builder">The <see cref="IFunctionsWorkerApplicationBuilder"/> to configure.</param>
         /// <returns>The same instance of the <see cref="IFunctionsWorkerApplicationBuilder"/> for chanining.</returns>
@@ -105,6 +105,26 @@ namespace Microsoft.Extensions.Hosting
                     var middleware = context.InstanceServices.GetRequiredService<T>();
 
                     return middleware.Invoke(context, next);
+                };
+            });
+
+            return builder;
+        }
+
+        /// <summary>
+        /// Configures the <see cref="IFunctionsWorkerApplicationBuilder"/> to use the provided inline middleware delegate.
+        /// </summary>
+        /// <param name="builder">The <see cref="IFunctionsWorkerApplicationBuilder"/> to configure.</param>
+        /// <param name="middleware">The middleware to add to the invocation pipeline.</param>
+        /// <returns>The same <see cref="IFunctionsWorkerApplicationBuilder"/> for chaining.</returns>
+        public static IFunctionsWorkerApplicationBuilder UseMiddleware(this IFunctionsWorkerApplicationBuilder builder, Func<FunctionContext, Func<Task>, Task> middleware)
+        {
+
+            builder.Use(next =>
+            {
+                return context =>
+                {
+                    return middleware(context, () => next.Invoke(context));
                 };
             });
 
